Put database into persistant volume.
Store the database in the /data directory instead of next to the executable.

/data can then be mapped to a host volume.

Also need to map in a hand-configured appsettings.json.

Downside of this is keeping the json file up to date with new versions;
unpacking a new mbbsemu archive will include an appsettings.json file -
if this is not deleted, it might fool someone into thinking it is used.